Refactored into auth and email service.
